--- crates/graph_search/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "agdb_graph_search"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-agdb_graph = { path = "../graph", version = "<=0.1.0" }
-agdb_bit_set = { path = "../bit_set", version = "<=0.1.0" }
