--- crates/graph_search/src/graph_search_from.rs
@@ -1,12 +1,0 @@
-use crate::graph_search::GraphSearch;
-use agdb_graph::GraphData;
-use agdb_graph::GraphImpl;
-
-impl<'a, Data> From<&'a GraphImpl<Data>> for GraphSearch<'a, Data>
-where
-    Data: GraphData,
-{
-    fn from(graph: &'a GraphImpl<Data>) -> Self {
-        GraphSearch { graph }
-    }
-}
