--- crates/graph_search/src/lib.rs
@@ -1,19 +1,0 @@
-mod breadth_first_search;
-mod breadth_first_search_reverse;
-mod depth_first_search;
-mod depth_first_search_reverse;
-mod graph_search;
-mod graph_search_from;
-mod path;
-mod path_search_handler;
-mod path_search_impl;
-mod search_control;
-mod search_handler;
-mod search_impl;
-mod search_index;
-mod search_iterator;
-
-pub use graph_search::GraphSearch;
-pub use path_search_handler::PathSearchHandler;
-pub use search_control::SearchControl;
-pub use search_handler::SearchHandler;
