--- crates/lib.rs
@@ -1,12 +1,0 @@
-mod db;
-mod query;
-mod query_error;
-mod query_result;
-mod transaction;
-
-pub use agdb_db_error::DbError;
-pub use db::Db;
-pub use query::Query;
-pub use query_error::QueryError;
-pub use query_result::QueryResult;
-pub use transaction::Transaction;
