--- crates/query.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct Query {}
