--- crates/tests/db_test.rs
@@ -1,13 +1,0 @@
-extern crate agdb;
-
-#[test]
-fn db_is_public_type() {
-    let _db = agdb::Db::default();
-}
-
-#[test]
-fn exec_takes_query_returns_query_result() {
-    let db = agdb::Db::default();
-    let query = agdb::Query::default();
-    let _result: Result<agdb::QueryResult, agdb::QueryError> = db.exec(query);
-}
