--- crates/tests/query_error_test.rs
@@ -1,6 +1,0 @@
-extern crate agdb;
-
-#[test]
-fn query_error_is_public_type() {
-    let _error = agdb::QueryError::default();
-}
