--- crates/tests/query_result_test.rs
@@ -1,6 +1,0 @@
-extern crate agdb;
-
-#[test]
-fn query_result_is_public_type() {
-    let _result = agdb::QueryResult::default();
-}
