--- crates/tests/query_test.rs
@@ -1,6 +1,0 @@
-extern crate agdb;
-
-#[test]
-fn query_is_public_type() {
-    let _query = agdb::Query::default();
-}
