+++ src/agdb/graph.rs
@@ -1,14 +1,14 @@
+pub mod graph_data;
+pub mod graph_impl;
 pub mod graph_index;
 pub mod storage_graph;
 
-mod graph_data;
 mod graph_data_memory;
 mod graph_data_storage;
 mod graph_data_storage_indexes;
 mod graph_edge;
 mod graph_edge_iterator;
 mod graph_edge_reverse_iterator;
-mod graph_impl;
 mod graph_node;
 mod graph_node_iterator;
