+++ src/agdb/graph_search/path.rs
@@ -1,4 +1,4 @@
-use agdb_graph::GraphIndex;
+use crate::graph::graph_index::GraphIndex;
 
 #[derive(Clone)]
 pub(crate) struct Path {
