+++ src/agdb/graph_search/path_search_handler.rs
@@ -1,4 +1,4 @@
-use agdb_graph::GraphIndex;
+use crate::graph::graph_index::GraphIndex;
 
 pub trait PathSearchHandler {
     fn process(&self, index: &GraphIndex, distance: &u64) -> u64;
