+++ src/agdb/graph_search/search_handler.rs
@@ -1,5 +1,5 @@
-use crate::search_control::SearchControl;
-use agdb_graph::GraphIndex;
+use super::search_control::SearchControl;
+use crate::graph::graph_index::GraphIndex;
 
 pub trait SearchHandler {
     fn process(&self, index: &GraphIndex, distance: &u64) -> SearchControl;
