+++ src/agdb/graph_search/search_index.rs
@@ -1,4 +1,4 @@
-use agdb_graph::GraphIndex;
+use crate::graph::graph_index::GraphIndex;
 
 pub(crate) struct SearchIndex {
     pub(crate) index: GraphIndex,
