+++ src/agdb/lib.rs
@@ -1,9 +1,20 @@
 mod collections;
 mod commands;
+mod db;
 mod db_error;
 mod graph;
+mod graph_search;
+mod query;
 mod storage;
+mod transaction;
 mod utilities;
 
 #[cfg(test)]
 mod test_utilities;
+
+pub use db::Db;
+pub use db_error::DbError;
+pub use query::Query;
+pub use query::QueryError;
+pub use query::QueryResult;
+pub use transaction::Transaction;
