+++ src/agdb/query.rs
@@ -1,0 +1,8 @@
+mod query_error;
+mod query_result;
+
+pub use query_error::QueryError;
+pub use query_result::QueryResult;
+
+#[derive(Default)]
+pub struct Query {}
