+++ src/agdb/query/query_error.rs
