+++ tests/query_error_test.rs
@@ -1,0 +1,6 @@
+use agdb::QueryError;
+
+#[test]
+fn query_error_is_public_type() {
+    let _error = QueryError::default();
+}
