+++ tests/query_result_test.rs
@@ -1,0 +1,6 @@
+use agdb::QueryResult;
+
+#[test]
+fn query_result_is_public_type() {
+    let _result = QueryResult::default();
+}
