+++ tests/query_test.rs
@@ -1,0 +1,6 @@
+use agdb::Query;
+
+#[test]
+fn query_is_public_type() {
+    let _query = Query::default();
+}
